🎨 Cielクラスを分割・フォーマット
