+++ cogs/error.py
@@ -2,7 +2,7 @@ from discord import Color, Embed, Interaction, app_commands
 from discord.app_commands import AppCommandError
 from discord.ext import commands
 
-from main import Ciel
+from ciel import Ciel
 
 
 class Error(commands.Cog):
